tlzapp/python/server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = "Facilitates functionality better provided by Python libraries"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.3",
    "pillow>=10.4.0",
    "protobuf>=5.28.2",
    "requests>=2.32.3",
    "sentencepiece>=0.2.0",
    "sqlite-vec>=0.1.3",
    "torch>=2.4.1",
    "torchvision>=0.21.0",
    "transformers>=4.51.2",
]
